backend/requirements.txt
Django==6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
python-decouple==3.8
django-cors-headers==4.7.0
psycopg2-binary==2.9.10
django-ratelimit==4.1.0
bleach==6.2.0
Pillow==11.1.0
google-generativeai==0.8.3
python-dotenv==1.0.0

# Redis and Celery
redis==5.0.1
django-redis==5.4.0
celery==5.3.4

# Reports and exports
reportlab==4.2.5
pandas==2.2.3
openpyxl==3.1.5
